move tokenizer trait to tokens module
