+++ llm-chain/src/text_splitter.rs
@@ -1,39 +1,5 @@
+use crate::tokens::{Tokenizer, TokenizerError};
 use std::cmp::max;
-use thiserror::Error;
-
-#[derive(Error, Debug)]
-pub enum TokenizerError {
-    #[error("Error tokenizing input text")]
-    TokenizationError,
-    #[error("Error stringifying tokens to text")]
-    ToStringError,
-    #[error("Error creating tokenizer")]
-    TokenizerCreationError,
-}
-
-pub trait Tokenizer<TokenType: Clone> {
-    /// Tokenizes a string.
-    ///
-    /// # Parameters
-    ///    
-    /// * `doc`: The string to tokenize.
-    ///
-    /// # Returns
-    ///
-    /// A `Result` containing a vector of tokens, or an error if there was a problem.
-    fn tokenize_str(&self, doc: &str) -> Result<Vec<TokenType>, TokenizerError>;
-
-    /// Converts a vector of tokens into a string.
-    ///
-    /// # Parameters
-    ///    
-    /// * `tokens`: The slice of tokens to convert.
-    ///
-    /// # Returns
-    ///
-    /// A `Result` containing a string, or an error if there was a problem.
-    fn to_string(&self, tokens: Vec<TokenType>) -> Result<String, TokenizerError>;
-}
 
 pub trait TextSplitter<TokenType>: Tokenizer<TokenType>
 where
